+++ app/src/main/res/layout/item_history.xml
@@ -24,7 +24,7 @@
             android:layout_height="100dp"
             card_view:layout_constraintStart_toStartOf="parent"
             card_view:layout_constraintTop_toTopOf="parent"
-            android:src="@tools:sample/avatars"
+            tools:src="@tools:sample/avatars"
             android:transitionName="image_history"
             />
