+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
     <item
-        android:id="@+id/navigation_beranda"
+        android:id="@+id/navigation_home"
         android:icon="@drawable/ic_home"
         android:title="@string/beranda" />
     <item
