+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -2,24 +2,26 @@
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/mobile_navigation">
+    android:id="@+id/mobile_navigation"
+    app:startDestination="@id/navigation_home"
+    >
 
     <fragment
         android:id="@+id/navigation_home"
         android:name="com.dicoding.aquaculture.view.main.HomeFragment"
-        android:label="Beranda"
+        android:label="@string/beranda"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/navigation_scan"
         android:name="com.dicoding.aquaculture.view.main.ScanFragment"
-        android:label="Scan"
+        android:label="@string/scan"
         tools:layout="@layout/fragment_scan" />
 
     <fragment
         android:id="@+id/navigation_profile"
         android:name="com.dicoding.aquaculture.view.main.ProfileFragment"
-        android:label="Profile"
+        android:label="@string/profile"
         tools:layout="@layout/fragment_profile" />
 
 </navigation>
